# [rails] Increase connection pool to total of threads